+++ lib/rangecoder.c
@@ -175,7 +175,7 @@ static inline void fill_prob( uint8 *img, uint32 *d, uint32 bits, uint32 a_bits,
 }
 */
 
-uint32  range_encoder(uint8 *img, uint32 *d, uint32 size, uint32 a_bits , uint32 q_bits, uint8 *buff, int *q)
+uint32  range_encoder(int16 *img, uint32 *d, uint32 size, uint32 a_bits , uint32 q_bits, uint8 *buff, int *q)
 /*! \fn uint32  range_encoder(uint8 *img, uint32 *distrib, const uint32 size, const uint8 bits)
 	\brief Range encoder.
     \param img	 	The pointer to encoding message data.
@@ -233,7 +233,7 @@ uint32  range_encoder(uint8 *img, uint32 *d, uint32 size, uint32 a_bits , uint32
 	return j;
 }
 
-uint32  range_decoder(uint8 *img, uint32 *d, uint32 size, uint32 a_bits , uint32 q_bits, uint8 *buff, int *q)
+uint32  range_decoder(int16 *img, uint32 *d, uint32 size, uint32 a_bits , uint32 q_bits, uint8 *buff, int *q)
 /*! \fn uint32  range_encoder(uint8 *img, uint32 *distrib, const uint32 size, const uint8 bits)
 	\brief Range decoder.
     \param img	 	The pointer to encoding message data.
@@ -290,7 +290,7 @@ uint32  range_decoder(uint8 *img, uint32 *d, uint32 size, uint32 a_bits , uint32
 	return j;
 }
 
-static uint32 make_distrib(uint8 *img, uint32 *d, uint32 size, uint32 a_bits ,uint32 q_bits, int *q)
+static uint32 make_distrib(int16 *img, uint32 *d, uint32 size, uint32 a_bits ,uint32 q_bits, int *q)
 /*! \fn uint32  range_encoder(uint8 *img, uint32 *distrib, const uint32 size, const uint8 bits)
 	\brief Range encoder.
     \param img	 	The pointer to encoding message data.
@@ -496,7 +496,7 @@ static inline uint32 get_cum_f1(uint32 out, uint32 *cu, uint32 i, uint32 j)
 	}
 }
 
-uint32  range_encoder1(uint8 *img, uint32 *d, uint32 size, uint32 a_bits , uint32 q_bits, uint8 *buff, int *q)
+uint32  range_encoder1(int16 *img, uint32 *d, uint32 size, uint32 a_bits , uint32 q_bits, uint8 *buff, int *q)
 /*! \fn uint32  range_encoder(uint8 *img, uint32 *distrib, const uint32 size, const uint8 bits)
 	\brief Range encoder.
     \param img	 	The pointer to encoding message data.
@@ -563,7 +563,7 @@ uint32  range_encoder1(uint8 *img, uint32 *d, uint32 size, uint32 a_bits , uint3
 	return j;
 }
 
-uint32  range_decoder1(uint8 *img, uint32 *d, uint32 size, uint32 a_bits , uint32 q_bits, uint8 *buff, int *q)
+uint32  range_decoder1(int16 *img, uint32 *d, uint32 size, uint32 a_bits , uint32 q_bits, uint8 *buff, int *q)
 /*! \fn uint32  range_encoder(uint8 *img, uint32 *distrib, const uint32 size, const uint8 bits)
 	\brief Range decoder.
     \param img	 	The pointer to encoding message data.
