+++ lib/rangecoder.h
@@ -19,11 +19,11 @@
 extern "C"{
 #endif /* __cplusplus */
 
-uint32  range_encoder (uint8 *img, uint32 *d, uint32 size, uint32 a_bits , uint32 q_bits, uint8 *buff, int *q);
-uint32  range_decoder (uint8 *img, uint32 *d, uint32 size, uint32 a_bits , uint32 q_bits, uint8 *buff, int *q);
-//uint32 range_encoder(uint8 *img, uint32 *d, const uint32 size, const uint32 bits , uint8 *buff);
-//uint32 range_decoder(uint8 *img, uint32 *d, const uint32 size, const uint32 bits , uint8 *buff);
-uint32  range_encoder1(uint8 *img, uint32 *d, uint32 size, uint32 a_bits , uint32 q_bits, uint8 *buff, int *q);
+uint32  range_encoder (int16 *img, uint32 *d, uint32 size, uint32 a_bits , uint32 q_bits, uint8 *buff, int *q);
+uint32  range_decoder (int16 *img, uint32 *d, uint32 size, uint32 a_bits , uint32 q_bits, uint8 *buff, int *q);
+//uint32 range_encoder(int16 *img, uint32 *d, const uint32 size, const uint32 bits , uint8 *buff);
+//uint32 range_decoder(int16 *img, uint32 *d, const uint32 size, const uint32 bits , uint8 *buff);
+uint32  range_encoder1(int16 *img, uint32 *d, uint32 size, uint32 a_bits , uint32 q_bits, uint8 *buff, int *q);
 
 #ifdef __cplusplus
 }
